Convert postinstall script to ts
